# contracts/api_v1.openapi.json
{
  "components": {
    "schemas": {
...
      "DecisionIn": {
        "properties": {
          "action": {
            "enum": [
              "dispatch_crew",
              "remote_check",
              "defer",
              "reject"
            ],
            "title": "Action",
            "type": "string"
          },
          "comment": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          },
          "reason_code": {
            "enum": [
              "false_alarm",
              "planned_works",
              "sensor_fault",
              "duplicate",
              "confirmed_by_readings",
              "confirmed_by_camera",
              "preventive",
              "no_crew",
              "await_data"
            ],
            "title": "Reason Code",
            "type": "string"
          }
        },
        "required": [
          "action",
          "reason_code"
        ],
        "title": "DecisionIn",
        "type": "object"
      },
...
      "ForecastCard": {
        "properties": {
          "asof": {
            "format": "date",
            "title": "Asof",
            "type": "string"
          },
          "calendar": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CalendarInfo"
              },
              {
                "type": "null"
              }
            ]
          },
          "case_key": {
            "title": "Case Key",
            "type": "string"
          },
          "channel": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ChannelRef"
              },
              {
                "type": "null"
              }
            ]
          },
          "coverage_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Coverage Note"
          },
          "data_status": {
            "enum": [
              "ok",
              "no_data",
              "stale"
            ],
            "title": "Data Status",
            "type": "string"
          },
          "decision": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DecisionOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "decisions": {
            "items": {
              "$ref": "#/components/schemas/DecisionOut"
            },
            "title": "Decisions",
            "type": "array"
          },
          "dynamics_30d": {
            "items": {
              "$ref": "#/components/schemas/DynamicsPoint"
            },
            "title": "Dynamics 30D",
            "type": "array"
          },
          "evidence": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evidence"
          },
          "factors": {
            "items": {
              "$ref": "#/components/schemas/FactorItem"
            },
            "title": "Factors",
            "type": "array"
          },
          "head": {
            "title": "Head",
            "type": "string"
          },
          "horizon_hours": {
            "title": "Horizon Hours",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "enum": [
              "alert",
              "weekly_recommendation"
            ],
            "title": "Kind",
            "type": "string"
          },
          "object": {
            "$ref": "#/components/schemas/ObjectRef"
          },
          "outcome_auto": {
            "anyOf": [
              {
                "enum": [
                  "hit",
                  "miss",
                  "unknown"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Outcome Auto"
          },
          "outcome_manual": {
            "anyOf": [
              {
                "enum": [
                  "confirmed_event",
                  "sensor_fault",
                  "normal_activation",
                  "no_event",
                  "unknown"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Outcome Manual"
          },
          "priority_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Priority Score"
          },
          "rank": {
            "title": "Rank",
            "type": "integer"
          },
          "recent_alarm_days_30": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recent Alarm Days 30"
          },
          "recent_alarm_days_7": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recent Alarm Days 7"
          },
          "risk": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Risk"
          },
          "scenario": {
            "enum": [
              "sensor_link",
              "equipment_diag",
              "guard_weekly"
            ],
            "title": "Scenario",
            "type": "string"
          },
          "scenario_title": {
            "title": "Scenario Title",
            "type": "string"
          },
          "score_type": {
            "enum": [
              "probability",
              "relative_priority"
            ],
            "title": "Score Type",
            "type": "string"
          },
          "source": {
            "enum": [
              "live",
              "emulated",
              "stub"
            ],
            "title": "Source",
            "type": "string"
          },
          "valid_from": {
            "format": "date-time",
            "title": "Valid From",
            "type": "string"
          },
          "valid_to": {
            "format": "date-time",
            "title": "Valid To",
            "type": "string"
          },
          "versions": {
            "items": {
              "$ref": "#/components/schemas/VersionItem"
            },
            "title": "Versions",
            "type": "array"
          },
          "work_order_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Work Order Id"
          }
        },
        "required": [
          "id",
          "kind",
          "scenario",
          "scenario_title",
          "head",
          "asof",
          "valid_from",
          "valid_to",
          "horizon_hours",
          "score_type",
          "rank",
          "object",
          "data_status",
          "source",
          "case_key"
        ],
        "title": "ForecastCard",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LoginIn": {
        "properties": {
          "login": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Login",
            "type": "string"
          },
          "password": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "login",
          "password"
        ],
        "title": "LoginIn",
        "type": "object"
      },
...
      "Page_ForecastItem_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ForecastItem"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "title": "Page Size",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "page_size"
        ],
        "title": "Page[ForecastItem]",
        "type": "object"
      },
      "Page_IngestBatchOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/IngestBatchOut"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "title": "Page Size",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "page_size"
        ],
        "title": "Page[IngestBatchOut]",
        "type": "object"
      },
...
      "RunDailyIn": {
        "properties": {
          "asof": {
            "format": "date",
            "title": "Asof",
            "type": "string"
          }
        },
        "required": [
          "asof"
        ],
        "title": "RunDailyIn",
        "type": "object"
      },
      "RunDailyOut": {
        "properties": {
          "asof": {
            "format": "date",
            "title": "Asof",
            "type": "string"
          },
          "forecasts_upserted": {
            "title": "Forecasts Upserted",
            "type": "integer"
          },
          "heads": {
            "additionalProperties": {
              "$ref": "#/components/schemas/HeadRunResult"
            },
            "title": "Heads",
            "type": "object"
          },
          "run_id": {
            "title": "Run Id",
            "type": "integer"
          },
          "work_orders_upserted": {
            "title": "Work Orders Upserted",
            "type": "integer"
          }
        },
        "required": [
          "run_id",
          "asof",
          "heads",
          "forecasts_upserted",
          "work_orders_upserted"
        ],
        "title": "RunDailyOut",
        "type": "object"
      },
...
      "UserOut": {
        "properties": {
          "login": {
            "title": "Login",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "permissions": {
            "items": {
              "type": "string"
            },
            "title": "Permissions",
            "type": "array"
          },
          "role": {
            "enum": [
              "dispatcher",
              "technician",
              "analyst",
              "manager",
              "admin",
              "integration"
            ],
            "title": "Role",
            "type": "string"
          }
        },
        "required": [
          "login",
          "name",
          "role",
          "permissions"
        ],
        "title": "UserOut",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Контракт C2 плана команды: backend → frontend. Прогнозы считает ML-сервис (C1); интерфейс читает только этот API.",
    "title": "Москоллектор — сервис диспетчера",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/admin/run-daily": {
      "post": {
        "operationId": "run_daily_api_v1_admin_run_daily_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunDailyIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunDailyOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Daily",
        "tags": [
          "admin"
        ]
      }
    },
...
    "/api/v1/auth/login": {
      "post": {
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Login",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/v1/forecasts": {
      "get": {
        "operationId": "list_forecasts_api_v1_forecasts_get",
        "parameters": [
          {
            "in": "query",
            "name": "scenario",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "sensor_link",
                    "equipment_diag",
                    "guard_weekly"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Scenario"
            }
          },
          {
            "in": "query",
            "name": "from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From"
            }
          },
          {
            "in": "query",
            "name": "to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To"
            }
          },
          {
            "in": "query",
            "name": "decision",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "none",
                    "any",
                    "dispatch_crew",
                    "remote_check",
                    "defer",
                    "reject"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Decision"
            }
          },
          {
            "in": "query",
            "name": "outcome",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "hit",
                    "miss",
                    "unknown"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Outcome"
            }
          },
          {
            "in": "query",
            "name": "obj",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Obj"
            }
          },
          {
            "in": "query",
            "name": "group_by",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "obj",
                    "case_key"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Group By"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 500,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_ForecastItem_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Forecasts",
        "tags": [
          "forecasts"
        ]
      }
    },
    "/api/v1/forecasts/{forecast_id}": {
      "get": {
        "operationId": "get_card_api_v1_forecasts__forecast_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "forecast_id",
            "required": true,
            "schema": {
              "title": "Forecast Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForecastCard"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Card",
        "tags": [
          "forecasts"
        ]
      }
    },
    "/api/v1/forecasts/{forecast_id}/decisions": {
      "post": {
        "operationId": "create_decision_api_v1_forecasts__forecast_id__decisions_post",
        "parameters": [
          {
            "in": "path",
            "name": "forecast_id",
            "required": true,
            "schema": {
              "title": "Forecast Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecisionIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Decision",
        "tags": [
          "forecasts"
        ]
      }
    },
    "/api/v1/forecasts/{forecast_id}/outcome": {
      "post": {
        "operationId": "set_outcome_api_v1_forecasts__forecast_id__outcome_post",
        "parameters": [
          {
            "in": "path",
            "name": "forecast_id",
            "required": true,
            "schema": {
              "title": "Forecast Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OutcomeIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OutcomeOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Outcome",
        "tags": [
          "forecasts"
        ]
      }
    },
...
    "/api/v1/ingest/batches": {
      "get": {
        "operationId": "list_batches_api_v1_ingest_batches_get",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 500,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_IngestBatchOut_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Batches",
        "tags": [
          "ingest"
        ]
      }
    },
    "/api/v1/ingest/day/{day}": {
      "delete": {
        "operationId": "reset_day_api_v1_ingest_day__day__delete",
        "parameters": [
          {
            "in": "path",
            "name": "day",
            "required": true,
            "schema": {
              "format": "date",
              "title": "Day",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResetDayOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reset Day",
        "tags": [
          "ingest"
        ]
      }
    },
    "/api/v1/ingest/events": {
      "post": {
        "description": "Пачка журнала СМВУ в JSON (до 5 000 строк) — так шлёт replay.py.",
        "operationId": "ingest_rows_api_v1_ingest_events_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "items": {
                  "$ref": "#/components/schemas/EventRowIn"
                },
                "title": "Rows",
                "type": "array"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestBatchOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ingest Rows",
        "tags": [
          "ingest"
        ]
      }
    },
    "/api/v1/ingest/events/upload": {
      "post": {
        "description": "Файл журнала СМВУ: CSV как журнал_событий_пример.csv или XLSX (ТЗ §7).",
        "operationId": "ingest_file_api_v1_ingest_events_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_ingest_file_api_v1_ingest_events_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestBatchOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ingest File",
        "tags": [
          "ingest"
        ]
      }
    },
    "/api/v1/ingest/ods-journal": {
      "post": {
        "operationId": "ingest_ods_api_v1_ingest_ods_journal_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "items": {
                  "$ref": "#/components/schemas/OdsRowIn"
                },
                "title": "Rows",
                "type": "array"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestBatchOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ingest Ods",
        "tags": [
          "ingest"
        ]
      }
    },
...
  }
}